adding explicit cast to u_int32_t for arc4random_uniform

+++ NSMutableArray+Shuffle.m
@@ -16,7 +16,7 @@
     for (NSUInteger i = 0; i < count; ++i) {
         // Select a random element between i and end of array to swap with.
         NSInteger nElements = count - i;
-        NSInteger n = arc4random_uniform(nElements) + i;
+        NSInteger n = arc4random_uniform((u_int32_t)nElements) + i;
         [self exchangeObjectAtIndex:i withObjectAtIndex:n];
     }
 }
